data validation with zod and react hook form

+++ src/actions/actions.ts
@@ -3,34 +3,52 @@
 import prisma from "@/lib/db";
 import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { Pet } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
 //it takes whatever we are putting in the form using action attribute and updating the db
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
+
+  //if not properly validated we do not want to intereact with the database
+  const validatedPet = petFormSchema.safeParse(pet);
+  if(!validatedPet.success){
+    return {
+      message:"Invalid pet data..",
+    };
+  }
 
   try {
     await prisma.pet.create({
       //creating new pet in the db
-      data: pet,
+      data: validatedPet.data
     });
   } catch (error) {
-    return {
+    return { 
       message: "couldn't add pets",
     };
   }
   revalidatePath("/app", "layout"); //we can instruct next js to re-render our layout
 }
 
-export async function editPet(petId: Pet['id'], newPetData :PetEssentials) {
+export async function editPet(petId: unknown, newPetData :unknown) {
   await sleep(1000);
+  
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+
+  if(!validatedPetId.success || !validatedPet.success){
+    return {
+      message:"Invalid pet data..",
+    };
+  }
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -40,12 +58,19 @@ export async function editPet(petId: Pet['id'], newPetData :PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId :Pet['id']) {
+export async function deletePet(petId :unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if(!validatedPetId.success){
+    return {
+      message:"Invalid pet data..",
+    };
+  }
   await sleep(1000);
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
